\name{wPerm-package}
\alias{wPerm-package}
\alias{wPerm}
\docType{package}

\title{
Permutation Tests
}

\description{
Supplies permutation-test alternatives to traditional hypothesis-test
procedures such as two-sample tests for means, medians, and standard deviations;
correlation tests; tests for homogeneity and independence; and more. Suitable for
general audiences, including individual and group users, introductory statistics
courses, and more advanced statistics courses that desire an introduction to
permutation tests.
}

\details{
\tabular{ll}{
Package: \tab wPerm\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2015-11-03\cr
License: \tab GPL (>= 2)\cr
}
}

\author{
Neil A. Weiss

Maintainer: Neil A. Weiss <neil.a.weiss@asu.edu>

}

\keyword{ package, datasets, graphs }

\concept{ permutation tests }
\concept{ randomization tests }
\concept{ simulation }
\concept{ inference }
